+++ src/Storage/Storage.Autorest/test/Set-AzStorageAdvancedPlatformMetric.Tests.ps1
@@ -15,7 +15,27 @@ if(($null -eq $TestName) -or ($TestName -contains 'Set-AzStorageAdvancedPlatform
 }
 
 Describe 'Set-AzStorageAdvancedPlatformMetric' {
-    It 'UpdateExpanded' -skip {
-        { throw [System.NotImplementedException] } | Should -Not -Throw
+    It 'CreateUpdateGetRemove-AdvancedPlatformMetric' {
+        $apt = Set-AzStorageAdvancedPlatformMetric -ResourceGroupName $env.ResourceGroupName -AccountName $env.AptAccount -RuleConfigFilterType AllContainersFilter  -Enabled
+        $apt.RuleConfigFilterType | Should -Be "AllContainersFilter"
+        $apt.RuleConfigFilterValue | Should -BeNullOrEmpty 
+        $apt.Enabled | Should -be $true
+
+        $apt = Set-AzStorageAdvancedPlatformMetric -ResourceGroupName $env.ResourceGroupName -AccountName $env.AptAccount -RuleConfigFilterType ContainerPrefixFilter -RuleConfigFilterValue "logs","data"  -Enabled 
+        $apt.RuleConfigFilterType | Should -Be "ContainerPrefixFilter"
+        $apt.RuleConfigFilterValue | Should -Be "logs","data"
+        $apt.Enabled | Should -be $true
+
+        $apt = Set-AzStorageAdvancedPlatformMetric -ResourceGroupName $env.ResourceGroupName -AccountName $env.AptAccount -RuleConfigFilterType ContainerListFilter -RuleConfigFilterValue "container1","container2","container3"
+        $apt.RuleConfigFilterType | Should -Be "ContainerListFilter"
+        $apt.RuleConfigFilterValue | Should -Be "container1","container2","container3"
+        $apt.Enabled | Should -be $false
+
+        $apt = Get-AzStorageAdvancedPlatformMetric -ResourceGroupName $env.ResourceGroupName -AccountName $env.AptAccount
+        $apt.RuleConfigFilterType | Should -Be "ContainerListFilter"
+        $apt.RuleConfigFilterValue | Should -Be "container1","container2","container3"
+        $apt.Enabled | Should -be $false
+
+         Remove-AzStorageAdvancedPlatformMetric -ResourceGroupName $env.ResourceGroupName -AccountName $env.AptAccount
     }
 }